pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mealie-mcp"
version = "0.1.0"
description = "Model Context Protocol server for the Mealie self-hosted recipe manager"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "eds3028" },
]
keywords = ["mcp", "mealie", "recipes", "llm", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
    "ruff>=0.5.0",
]

[project.scripts]
mealie-mcp = "mealie_mcp.__main__:main"

[project.urls]
Homepage = "https://github.com/eds3028/mealie-mcp"
Repository = "https://github.com/eds3028/mealie-mcp"
Issues = "https://github.com/eds3028/mealie-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/mealie_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
